refactor: backend refactored
