--- Server/MessageAppServer/Contracts/ChangeEmailResponse.cs
@@ -1,3 +1,0 @@
-﻿namespace Server.Contracts;
-
-public record ChangeEmailResponse(string Email, string UserName);
