--- Server/MessageAppServer/Contracts/ChangeUserPasswordResponse.cs
@@ -1,3 +1,0 @@
-﻿namespace Server.Contracts;
-
-public record ChangeUserPasswordResponse(string Email, string UserName);
