--- Server/MessageAppServer/Contracts/RegistrationResponse.cs
@@ -1,3 +1,0 @@
-﻿namespace Server.Contracts;
-
-public record RegistrationResponse(string Email, string UserName);
