+++ Server/MessageAppServer/Database/RoomsContext.cs
@@ -3,8 +3,7 @@ using Server.Model.Chat;
 
 namespace Server.Database;
 
-public class RoomsContext : DbContext
+public class RoomsContext(DbContextOptions<RoomsContext> options, DbSet<Room> rooms) : DbContext(options)
 {
-    public DbSet<Room> Rooms { get; set; }
-    public RoomsContext(DbContextOptions<RoomsContext> options) : base(options) { }
+    public DbSet<Room> Rooms { get; set; } = rooms;
 }
