+++ Server/MessageAppServer/Requests/ChangeEmailRequest.cs
@@ -1,5 +1,5 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Server.Contracts;
+namespace Server.Requests;
 
 public record ChangeEmailRequest([Required]string OldEmail, [Required]string NewEmail, [Required]string Token);
