+++ Server/MessageAppServer/Requests/ChangeUserPasswordRequest.cs
@@ -1,5 +1,5 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Server.Contracts;
+namespace Server.Requests;
 
 public record ChangeUserPasswordRequest([Required]string Email, [Required]string OldPassword, [Required]string NewPassword);
