+++ Server/MessageAppServer/Requests/MessageRequest.cs
@@ -1,5 +1,5 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Server.Services.Chat.MessageService;
+namespace Server.Requests;
 
 public record MessageRequest([Required]string RoomId, [Required]string UserName, [Required]string Message);
