+++ Server/MessageAppServer/Requests/RegistrationRequest.cs
@@ -1,5 +1,5 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Server.Contracts;
+namespace Server.Requests;
 
 public record RegistrationRequest([Required]string Email, [Required]string Username, [Required]string Password, [Required]string PhoneNumber, string Image);
