+++ Server/MessageAppServer/Requests/RoomRequest.cs
@@ -1,5 +1,5 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Server.Services.Chat;
+namespace Server.Requests;
 
 public record RoomRequest([Required]string RoomName, [Required]string Password);
