+++ Server/MessageAppServer/Responses/AuthResponse.cs
@@ -1,3 +1,3 @@
-﻿namespace Server.Services.Authentication;
+﻿namespace Server.Responses;
 
 public record AuthResponse(string Email, string Username, string Token);
