+++ Server/MessageAppServer/Responses/EmailUsernameResponse.cs
@@ -1,0 +1,3 @@
+﻿namespace Server.Responses;
+
+public record EmailUsernameResponse(string Email, string UserName);
