+++ Server/MessageAppServer/Responses/GetEmailForVerificationResponse.cs
@@ -1,3 +1,3 @@
-﻿namespace Server.Contracts;
+﻿namespace Server.Responses;
 
 public record GetEmailForVerificationResponse(string Message);
