+++ Server/MessageAppServer/Responses/RoomResponse.cs
@@ -1,3 +1,3 @@
-﻿namespace Server.Services.Chat;
+﻿namespace Server.Responses;
 
 public record RoomResponse(bool Success, string RoomId, string RoomName) { }
