+++ Server/MessageAppServer/Responses/UserResponse.cs
@@ -1,3 +1,3 @@
-﻿namespace Server.Contracts;
+﻿namespace Server.Responses;
 
 public record UserResponse(string? Email, bool TwoFactorEnabled);
