+++ Server/MessageAppServer/Services/Chat/MessageService/IMessageService.cs
@@ -1,4 +1,6 @@
 ﻿using Server.Model.Chat;
+using Server.Requests;
+using Server.Responses;
 
 namespace Server.Services.Chat.MessageService;
