+++ Server/MessageAppServer/Services/Chat/MessageService/MessageService.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Server.Database;
 using Server.Model.Chat;
+using Server.Requests;
+using Server.Responses;
 
 namespace Server.Services.Chat.MessageService;
