+++ Server/MessageAppServer/Services/EmailSender/EmailSenderCodeGenerator.cs
@@ -12,9 +12,9 @@ public static class EmailSenderCodeGenerator
         var rnd = new Random();
         var token = new char[19];
 
-        for (int i = 0; i < 19; i++)
+        for (var i = 0; i < 19; i++)
         {
-            if (i > 0 && i == 4 || i > 0 && i == 9 || i > 0 && i == 14)
+            if (i is > 0 and 4 or > 0 and 9 or > 0 and 14)
             {
                 token[i] = '-';
             }
@@ -35,10 +35,10 @@ public static class EmailSenderCodeGenerator
 
     public static bool ExamineIfTheCodeWasOk(string email, string verifyCode)
     {
-        _verificationCodes.TryGetValue(email, out string value);
+        _verificationCodes.TryGetValue(email, out var value);
         Console.WriteLine(value);
 
-        return _verificationCodes.ContainsKey(email) ? _verificationCodes[email] == verifyCode : false;
+        return _verificationCodes.TryGetValue(email, out var code) && code == verifyCode;
     }
 
     public static void RemoveVerificationCode(string email)
